+++ app/neo/artifact/DeleteConfirm.tsx
