+++ app/routes/forms._index/route.tsx
@@ -1,11 +1,9 @@
-import { Button, Flex, SearchInput, Spinner } from '@axonivy/ui-components';
-import { IvyIcons } from '@axonivy/ui-icons';
+import { Flex, SearchInput, Spinner } from '@axonivy/ui-components';
 import type { MetaFunction } from '@remix-run/node';
 import { useState } from 'react';
 import { Form, useCreateForm, useDeleteForm, useForms } from '~/data/form-api';
 import { ProjectIdentifier } from '~/data/project-api';
-import { ArtifactCard, cardLinks } from '~/neo/card/ArtifactCard';
-import { useNewArtifactDialog } from '~/neo/dialog/useNewArtifactDialog';
+import { ArtifactCard, NewArtifactCard, cardLinks } from '~/neo/artifact/ArtifactCard';
 import FormPreviewSVG from './form-preview.svg?react';
 import { editorIcon, editorId, useEditors } from '~/neo/useEditors';
 
@@ -20,17 +18,22 @@ export default function Index() {
   const { data, isPending } = useForms();
   const forms = data?.filter(form => form.name.toLocaleLowerCase().includes(search.toLocaleLowerCase())) ?? [];
   const { createForm } = useCreateForm();
-  const { open } = useNewArtifactDialog();
   const create = (name: string, namespace: string, project?: ProjectIdentifier) => createForm({ name, namespace, type: 'Form', project });
   return (
     <Flex direction='column' gap={4} style={{ padding: 30, height: 'calc(100% - 60px)', overflowY: 'auto' }}>
-      <Flex direction='row' alignItems='center' justifyContent='space-between'>
-        <span style={{ fontWeight: 600, fontSize: 16 }}>Forms</span>
-        <Button icon={IvyIcons.Plus} size='large' onClick={() => open({ create, defaultName: 'MyNewForm', title: 'Create new Form' })} />
-      </Flex>
+      <span style={{ fontWeight: 600, fontSize: 16 }}>Forms</span>
       <SearchInput value={search} onChange={setSearch} />
       <Flex gap={4} style={{ flexWrap: 'wrap' }}>
-        {isPending ? <Spinner size='small' /> : forms.map(form => <FormCard key={form.path ?? form.name} {...form} />)}
+        {isPending ? (
+          <Spinner size='small' />
+        ) : (
+          <>
+            <NewArtifactCard create={create} title='Create new Form' defaultName='MyNewForm' />
+            {forms.map(form => (
+              <FormCard key={form.path ?? form.name} {...form} />
+            ))}
+          </>
+        )}
       </Flex>
     </Flex>
   );
