+++ app/routes/processes._index/route.tsx
@@ -1,11 +1,9 @@
-import { Flex, Button, SearchInput, Spinner } from '@axonivy/ui-components';
-import { IvyIcons } from '@axonivy/ui-icons';
+import { Flex, SearchInput, Spinner } from '@axonivy/ui-components';
 import { MetaFunction } from '@remix-run/node';
 import { useState } from 'react';
 import { useProcesses, useDeleteProcess, useCreateProcess, Process } from '~/data/process-api';
 import { ProjectIdentifier } from '~/data/project-api';
-import { ArtifactCard, cardLinks } from '~/neo/card/ArtifactCard';
-import { useNewArtifactDialog } from '~/neo/dialog/useNewArtifactDialog';
+import { ArtifactCard, NewArtifactCard, cardLinks } from '~/neo/artifact/ArtifactCard';
 import { useEditors, editorId, editorIcon } from '~/neo/useEditors';
 import ProcessPreviewSVG from './process-preview.svg?react';
 
@@ -20,22 +18,23 @@ export default function Index() {
   const { data, isPending } = useProcesses();
   const processes = data?.filter(proc => proc.name.toLocaleLowerCase().includes(search.toLocaleLowerCase())) ?? [];
   const { createProcess } = useCreateProcess();
-  const { open } = useNewArtifactDialog();
   const create = (name: string, namespace: string, project?: ProjectIdentifier) =>
     createProcess({ name, namespace, kind: 'Business Process', project });
   return (
     <Flex direction='column' gap={4} style={{ padding: 30, height: 'calc(100% - 60px)', overflowY: 'auto' }}>
-      <Flex direction='row' alignItems='center' justifyContent='space-between'>
-        <span style={{ fontWeight: 600, fontSize: 16 }}>Processes</span>
-        <Button
-          icon={IvyIcons.Plus}
-          size='large'
-          onClick={() => open({ create, title: 'Create new Process', defaultName: 'MyNewProcess' })}
-        />
-      </Flex>
+      <span style={{ fontWeight: 600, fontSize: 16 }}>Processes</span>
       <SearchInput value={search} onChange={setSearch} />
       <Flex gap={4} style={{ flexWrap: 'wrap' }}>
-        {isPending ? <Spinner size='small' /> : processes.map(process => <ProcessCard key={process.path ?? process.name} {...process} />)}
+        {isPending ? (
+          <Spinner size='small' />
+        ) : (
+          <>
+            <NewArtifactCard create={create} title='Create new Process' defaultName='MyNewProcess' />
+            {processes.map(process => (
+              <ProcessCard key={process.path ?? process.name} {...process} />
+            ))}
+          </>
+        )}
       </Flex>
     </Flex>
   );
